requirements-docker.txt
alembic==1.13.2
blinker==1.8.2
cement==2.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
docutils==0.16
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.5
MarkupSafe==2.1.5
pathspec==0.10.1
psycopg2==2.9.9
pyasn1==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.2
semantic-version==2.8.5
six==1.16.0
SQLAlchemy==2.0.33
termcolor==1.1.0
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.1.9
Werkzeug==3.0.4
